+++ internal/store/indexer.go
@@ -17,8 +17,27 @@ type ObjectReference struct {
 	Referent client.Object
 }
 
-// key format:
-// group/version,Kind=kind/namespace/name:group/version,Kind=kind/namespace/name.
+// CacheIndexers implements a reference cache to store reference relationship between k8s objects
+// provided by cache.Indexer. It could do CRUD on reference records when referrer and referent are
+// both provided. It can also list reference records by referrer or by referent.
+type CacheIndexers struct {
+	indexer cache.Indexer
+}
+
+func NewCacheIndexers() CacheIndexers {
+	return CacheIndexers{
+		indexer: cache.NewIndexer(ObjectReferenceKeyFunc,
+			cache.Indexers{
+				IndexNameReferrer: ObjectReferenceIndexerReferrer,
+				IndexNameReferent: ObjectReferenceIndexerReferent,
+			}),
+	}
+}
+
+// ObjectReferenceKeyFunc is the function to trasfer the reference relataioship
+// between k8s objects. The key is transferred to the following format:
+// group/version,Kind=kind/namespace/name:group/version,Kind=kind/namespace/name
+// The part before : is from referrer, and after : is from referent.
 func ObjectReferenceKeyFunc(obj interface{}) (string, error) {
 	ref, ok := obj.(*ObjectReference)
 	if !ok {
@@ -36,6 +55,9 @@ func ObjectReferenceKeyFunc(obj interface{}) (string, error) {
 	return referrerKey + ":" + referentKey, nil
 }
 
+// ObjectReferenceIndexerReferrer is the index function to index by the referrer,
+// which returns index in the follwig format from referrer:
+// group/version,Kind=kind/namespace/name.
 func ObjectReferenceIndexerReferrer(obj interface{}) ([]string, error) {
 	ref, ok := obj.(*ObjectReference)
 	if !ok {
@@ -47,6 +69,9 @@ func ObjectReferenceIndexerReferrer(obj interface{}) ([]string, error) {
 	)}, nil
 }
 
+// ObjectReferenceIndexerReferent is the index function to index by the referent,
+// which returns index in the following format from referent:
+// group/version,Kind=kind/namespace/name.
 func ObjectReferenceIndexerReferent(obj interface{}) ([]string, error) {
 	ref, ok := obj.(*ObjectReference)
 	if !ok {
@@ -58,28 +83,18 @@ func ObjectReferenceIndexerReferent(obj interface{}) ([]string, error) {
 	)}, nil
 }
 
-type CacheIndexers struct {
-	indexer cache.Indexer
-}
-
-func NewCacheIndexers() CacheIndexers {
-	return CacheIndexers{
-		indexer: cache.NewIndexer(ObjectReferenceKeyFunc,
-			cache.Indexers{
-				IndexNameReferrer: ObjectReferenceIndexerReferrer,
-				IndexNameReferent: ObjectReferenceIndexerReferent,
-			}),
-	}
-}
-
+// SetReference stores a reference record in the cache.
 func (c CacheIndexers) SetReference(ref *ObjectReference) error {
 	return c.indexer.Add(ref)
 }
 
-func (c CacheIndexers) RemoveReference(ref *ObjectReference) error {
+// DeleteReference deletes a reference record (with given key of referrer and referent).
+func (c CacheIndexers) DeleteReference(ref *ObjectReference) error {
 	return c.indexer.Delete(ref)
 }
 
+// ObjectReferred returns true if an object is referenced (being the referent)
+// in at least one reference record.
 func (c CacheIndexers) ObjectReferred(obj client.Object) (bool, error) {
 	key := fmt.Sprintf("%s/%s/%s",
 		obj.GetObjectKind().GroupVersionKind().String(),
@@ -94,6 +109,27 @@ func (c CacheIndexers) ObjectReferred(obj client.Object) (bool, error) {
 	return len(refs) != 0, nil
 }
 
+// ListReferencesByReferrer list all reference records where referrer has the same key
+// (GroupVersionKind+NamespacedName, that means the same k8s object).
+func (c CacheIndexers) ListReferencesByReferrer(referrer client.Object) ([]*ObjectReference, error) {
+	key := fmt.Sprintf("%s/%s/%s",
+		referrer.GetObjectKind().GroupVersionKind().String(),
+		referrer.GetNamespace(), referrer.GetName(),
+	)
+	refList, err := c.indexer.ByIndex(IndexNameReferrer, key)
+	if err != nil {
+		return nil, err
+	}
+	returnRefList := []*ObjectReference{}
+	for _, ref := range refList {
+		returnRefList = append(returnRefList, ref.(*ObjectReference))
+	}
+	return returnRefList, nil
+}
+
+// DeleteReferencesByReferrer deletes all reference records where referrer has the same key
+// (GroupVersionKind+NamespacedName, that means the same k8s object).
+// called when a k8s object deleted in cluster, or when we do not care about it anymore.
 func (c CacheIndexers) DeleteReferencesByReferrer(referrer client.Object) error {
 	key := fmt.Sprintf("%s/%s/%s",
 		referrer.GetObjectKind().GroupVersionKind().String(),